0.5.4 (#68)
* 0.5.4

Update package version and dependencies, remove className prop from DialogPortal

Upgraded the package version from 0.5.3 to 0.5.4. Bumped up the versions of several dependencies in package.json and package-lock.json. In Dialog.tsx, the `className` prop was removed from DialogPortal. This would enhance overall package performance and compatibility with its dependencies. The removal of `className` prop was likely because it was unused or the functionality was achieved by another means.

* Remove CI-specific screenshot on failure

The screenshotOnRunFailure option was removed from the Cypress configuration as it was causing unnecessary clutter and confusion in the CI environment. Now, screenshots will only be taken in local environments, as intended.