Update UI components for better responsiveness and dark mode support
Several updates were made across the Dialog, Modal, Dropdown and Select UI components primarily to improve appearance and behavior in different contexts specifically dark mode. Redundancies were removed and code was optimized where needed. The updates also include more flexibility in positioning and alignment for the Dropdown component.

Additionally, a minor version number increase in package.json is reflected.

The goal was to enhance the adaptability of components in varied layout and theme settings, as well as achieve a cleaner, more efficient code.