0.6.4 (#72)
* 0.6.4

Remove CallToActionButton and EmailNavBar, upgrade packages

Deleted CallToActionButton and EmailNavbar components due to redundancy. Updated the package version in package.json, and added new package @react-email/components. Also, updated Ubuntu version to latest in the Github workflow build.

The deleted components were not adding additional functionality and were removed to clean up the codebase. The packages were updated for better performance and to remove any potential vulnerabilities. The Ubuntu version was updated to the latest version to ensure compatibility and performance for the build process.

* Add visual feedback for currently selected theme

A CheckCircleIcon has been added to ProfileDropdown.tsx to provide visual feedback to the user about the currently selected theme. Now, whenever a theme - light, dark, or system - is selected, a checkmark will appear next to it. This change is intended to improve the user experience by providing visual confirmation of the active theme.

* Add step to stop Supabase in GitHub workflow

The change adds a new step in the Github CI/CD workflow to stop the Supabase server once the workflow run is complete. This ensures that the test environment is cleaned up after each run, freeing up resources for other jobs

* Ensure Supabase always stops after workflow

An `if: always()` statement was added to the 'Stop Supabase' step in .github/workflows/build.yml. This ensures that the Supabase service always stops after the workflow runs, regardless of the outcome of previous steps. This is essential to prevent unwanted resource usage.
* Update error handling in Cypress commands

Expanded the error handling in cypress/support/commands.ts to also skip "ResizeObserver loop limit exceeded" errors, preventing unnecessary test failures. This adjustment ensures more accurate test results by ignoring irrelevant errors.