0.6.6 (#74)
* 0.6.6

Refactor user session data handling

Updated the way the user session data is handled across various components. The auth and user details are now declared explicitly in the UserSession type as separate fields. This pattern aims for more comprehensive typing of the user session and for a clearer understanding of the session data structure. The changes have been propagated through various parts of the code improving readability and data manipulation simplicity. Also, the package version has been bumped.

* Add redirect check to Next action detection

Added a new constant `NEXT_ACTION_REDIRECT_HEADER` and modified the `isNextAction` function to also check if the request is not a redirect from a Next action. This change ensures that the middleware handles redirects from Next actions correctly.