0.6.12 (#82)
* 0.6.12

Update and optimize package dependencies

A few dependencies in package.json and package-lock.json were updated to the latest versions for optimal performance and the removal of potential security risks. The project version was incremented from 0.6.11 to 0.6.12. Key updates included 'edge-csrf' from ^1.0.5-rc1 to ^1.0.5, 'pino' from ^8.15.5 to ^8.15.6, and 'pino-pretty' from 10.2.0 to 10.2.3. In addition, 'react-hot-toast' was removed and replaced with 'sonner' for better toasting functionalities. The changes also involved updating references to 'react-hot-toast' in the codebase to 'sonner'. These updates aim to maintain the project's relevance with the current JavaScript ecosystem and improve the overall quality of the application.

* Refactor Modal usage and improve semantic HTML representation

This commit contains a major overhaul of how modals are being utilized and invoked across several components. Previously, modals were invoked and managed using separate state variables which added unnecessary complexity and bloating of the component's state. With this update, we've moved to a more declarative and streamlined approach. Now, each Modal is paired directly with the triggering element. This makes modal handling more intuitive and reduces the necessity of managing modal states independently, which in turn results in cleaner, more maintainable code.

Additionally, it includes improvements in the semantic HTML representation across different components for better accessibility and interpretability of the user interface. Changes were executed in several files, including OrganizationsSelector.tsx, MultiFactorAuthenticationSettings.tsx, UpdatePhoneNumberForm.tsx, and others. This helps in improving the overall user experience and readability of the application code.

Also, relative import paths have been updated to absolute import paths improving readability and maintainability of the codebase.