Use edge runtime for some routes (#83)
* Use edge runtime for some routes

The method of importing translation files in `i18n.server.tsx` has been updated. Instead of using the 'fs/promises' module to read files, the 'import' function is now used to dynamically import translation files. This change simplifies the code and provides more straightforward error handling. It should improve the maintainability of the code and allow for easier debugging in the future.

* Use cookie to set language in I18nProvider"

This commit adds functionality to get and use the language preference from a 'lang' cookie. If the language is not specified, it utilizes the 'lang' cookie. We've added a new function, 'getLanguageFromCookie', to fetch the language preference. This change aims to improve user experience by remembering their language preference across sessions.

* Refactor user session handling and improve loading states

This commit refactors the way user sessions are handled in the SiteHeaderSessionProvider by passing session data as props. The changes help to avoid unnecessary re-rendering of the component when there's a change in the data.

The use of the `loadUserData` function in `SiteLayout` and `NotFoundPage` ensures that the data required for the user session is loaded correctly. This change promotes more efficiency in the application.

The loading component has also been enhanced to include `TopLoadingBarIndicator` for better user experience during data loading.

Lastly, the redundant 'language' property is removed from the configuration file - this is determined by 'locale' and doesn't need to be explicitly set.

* Remove deprecated `i18n` property from NavigationItem

The deprecated `i18n` property was removed from the NavigationItem component and its usage was refactored. Instead, the `label` property is now used exclusively for defining the displayed text of a navigation item, simplifying the interfaces and use of the component. In addition, the font weight of navigation items on bigger screens was changed from "font-semibold" to "font-medium" in the class builder function. This change is aimed at making the UI more consistent and visually appealing.