Update package version and refactor not-found.tsx (#84)
* Update package version and refactor not-found.tsx

The package version was updated to 0.6.14 in the package.json file. The not-found.tsx file was also modified, specifically refactoring the way user data is loaded and passed on to the SiteHeaderSessionProvider component. Instead of directly passing the session object, we are now unwrapping it from the data object that is returned from the loadData function. This change was made to make the data handling more consistent with the rest of our codebase.

* Update UI components and improve display of logs during development

A number of changes have been made in this commit, primarily to improve UI styling and app functionality.

Firstly, padding and icon sizes have been updated in 'Select.tsx' for enhanced visual appeal. The `organizationsSelector.tsx` & `CreateOrganizationModal.tsx` have been cleaned for better readability.

Changes in `package.json` and `logger.ts` improve the clarity of console logs in developer mode. This was achieved by combining all logs by adding `pino-pretty` as a tool to prettify the logs in the 'dev' script.

The `NotFoundPage` has been wrapped with the `I18nProvider` to translate the text on the 404 page.

Moreover, there were changes in other files mainly styling tweaks to enhance user experience and a minor bug fix by removing a redundant `minLength` rule.