0.6.17 (#89)
* 0.6.17

Replace temporary redirects by permanent ones and update third-party libraries

* Update Sentry initialization and enhance VerificationCodeInput

In the initializeBrowserSentry.ts, Sentry initialization was updated to check if DSN is provided earlier in the function. This avoids unnecessary import of modules if DSN is not provided, especially in non-production environments.

In VerificationCodeInput.tsx, a feature to handle paste and backspace operations was added. The paste operation helps in pasting the verification code directly from a clipboard if it's numeric. The backspace operation helps to move to the previous input box when backspace is pressed, thereby improving the user experience.

* "Update verification for non-admin user access"

This commit changes the verification in the test for non-admin user access. Previously, the test was checking if the user is redirected to the homepage. Given the recent changes in the application where a non-admin user trying to access an admin page is redirected to a 404 page, the test is updated to check for a 404 status code.

* Add server session refresh action after profile update

A new action to refresh the user session on the server is introduced, which is now called after updating the user profile. This has been implemented to ensure changes made to user's profile are accurately reflected in their session data. The refreshSessionAction function is imported into the component to update profile data and is called upon in the update functions.

* Refactor loading indicators and improve app redirects

The loading indicators across the application were simplified and contactualized. Now the `GlobalLoadingIndicator` component is used directly instead of wrapped in additional component functions. This eliminates unnecessary complexity and improves maintainability.

The loading message has been made more descriptive in `common.json`, and `TopLoadingBarIndicator` was removed from most loading components since `GlobalLoadingIndicator` already includes it.

Also, changed the `permanentRedirect` function to the `redirect` function across the application to ensures more consistent behaviour for url change operations.

The spacing and text format for `PageLoadingIndicator` were adjusted, and an unnecessary callback function in `UpdateProfileFormContainer.tsx` was removed to simplify code.

All changes are made following the best practices to simplify the code structure and improve user experiences while maintaining all existing functions.

* Update dependency versions for better compatibility and performance

Updated the versions of `nanoid` and `recharts` to newer versions(`nanoid` from 5.0.1 to 5.0.2 and `recharts` from 2.8.0 to 2.9.0) and reflected these changes in `package-lock.json`. Also, removed dependency on `css-unit-converter` and `reduce-css-calc` for optimized performance.