Update error handling logic in auth callback route (#96)
Refactor error handling in the callback of the authorization feature. An error function has been introduced for better reusability and readability. Also, handling of user invite acceptance error has been improved.

Additionally, the version of the application is updated in the package-lock file.

This reshuffling of the logic helps in better debugging in the future, and keeping the code clean.