0.7.7 (#98)
Update package version and remove unused dependencies

The package version was updated from 0.7.6 to 0.7.7 to keep track of the changes. Unused dependencies of "@react-email" were also removed to clean up the project and ensure that only the necessary packages are being used. "@supabase/auth-helpers-nextjs" was updated from "^0.8.1"  to "^0.8.2" to use the latest version and incorporate any updates or bug fixes from the package.